+++ packages/client/src/index.ts
@@ -1,2 +1,2 @@
-export { createRelay } from "./Document";
+export { createWeavo } from "./Document";
 export type { TextChange } from "./types";