+++ packages/transport/CHANGELOG.md
@@ -1,4 +1,4 @@
-# @relay/transport
+# @weavo/transport
 
 ## 1.0.0
 
@@ -9,5 +9,5 @@
 ### Patch Changes
 
 - Updated dependencies [b5cf20f]
-  - @relay/code@1.0.0
-  - @relay/sync@1.0.0
+  - @weavo/code@1.0.0
+  - @weavo/sync@1.0.0